pyproject.toml
[project]
name = "i"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cpuinfo"
]
